Consolidate NAICS strategies and clean up utils __init__.py re-exports
Merge 4 simple NAICS strategy classes (OriginalData, TopicCode, AgencyDefaults,
SectorFallback) into single simple_strategies.py file. Old files become 3-line
re-export shims for backward compatibility. Strip 38 unused re-exports from
sbir_etl/utils/__init__.py — none were imported from the package level.

~415 net lines removed.
