+++ sbir_etl/enrichers/naics/fiscal/strategies/__init__.py
@@ -2,14 +2,15 @@
 
 from __future__ import annotations
 
-from .agency_defaults import AgencyDefaultsStrategy
 from .base import EnrichmentStrategy, NAICSEnrichmentResult
-from .original_data import OriginalDataStrategy
-from .sector_fallback import SectorFallbackStrategy
+from .simple_strategies import (
+    AgencyDefaultsStrategy,
+    OriginalDataStrategy,
+    SectorFallbackStrategy,
+    TopicCodeStrategy,
+)
 from .text_inference import TextInferenceStrategy
-from .topic_code import TopicCodeStrategy
 from .usaspending_dataframe import USAspendingDataFrameStrategy
-
 
 __all__ = [
     "EnrichmentStrategy",
